Backend Swtiched to Node JS/Express from Django
Application Approach changed to focus on transfer students at Community Colleges transferring to CalPoly.

This approach is easier to do and we will do the webscraping beforehand to retrieve the class data that will be stored in a database.

Due to this we are using node js and express to make the route handling logic, database implementation, and user authentication easier.