Saved Schedule Page implemented that loads in data  (#14)
* Create Schedule (Save) to MongoDB database

User can now click a button and save their schedule in mongoDB to have the following object:
{
  "_id": {
    "$oid": "66478294ab10de31a754b768"
  },
  "userId": "VaXsgwkdzkWwgu4j4o0ZFMXFpan1",
  "schedule": [
    {
      "id": 0,
      "term": "Summer",
      "year": 2021,
      "courses": [
        {
          "id": 0,
          "courseCode": "CSC123",
          "units": "4.00"
        },
        {
          "id": 1,
          "courseCode": "ACCTG4A",
          "units": "4.00"
        },
        {
          "id": 2,
          "courseCode": "",
          "units": ""
        },
        {
          "id": 3,
          "courseCode": "",
          "units": ""
        }
      ]
    },
    {
      "id": 1,
      "term": "Fall",
      "year": 2021,
      "courses": [
        {
          "id": 10,
          "courseCode": "CSC101",
          "units": "4.00"
        },
        {
          "id": 11,
          "courseCode": "",
          "units": ""
        },
        {
          "id": 12,
          "courseCode": "",
          "units": ""
        },
        {
          "id": 13,
          "courseCode": "",
          "units": ""
        }
      ]
    },
...
  ],
  "params": {
    "year": "2024",
    "ccc": "Clovis Community College",
    "cccCode": "CLOVIS",
    "college": "California Polytechnic University, San Luis Obispo",
    "major": "COMPUTER SCIENCE, B.S."
  }
}

* Create, Delete, and Reading Schedules working

Create: Entire UI for create schedule is working.
Delete: Backend logic finished. Frontend is half finished where users can delete a recently saved schedule.
Read: Backend logic finished. Front end UI will be created in the next commit to allow users to choose to view, edit, or delete the scheudle in the list. For this step, I will list the schedules as cards in a new page "/savedSchedules"

* ScheduleCard Implemented

* scheduleCard now contains info from Database

Loading in the data for the saved schedules based on the userId and passing it to the scheduleList to map out the scheduleCards with the database information.

* Fixed small bug in UserMenu

Page wouldn't load if user wasn't signed in. Fixed this bug by allowing the path to "savedSchedules" to be accessed only if currentUser existed (user was logged in)